More UI updates to Filter Panel
git-svn-id: https://svn2.assembla.com/svn/savant/MedSavant/branches/1.0.1@3040 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ MedSavantClient/src/org/ut/biolab/medsavant/view/component/ProgressPanel.java
@@ -27,8 +27,8 @@ public class ProgressPanel extends JPanel implements ComponentListener {
     private boolean animating = false;
     private int speed = 300;
     private int maxValue = 100;
-    private int currentValue;
-    private int lastValue;
+    private long currentValue;
+    private long lastValue;
     private double animatingValue;
     private boolean drawLastIndicator = false;
     private boolean drawShadow = true;
@@ -131,7 +131,7 @@ public class ProgressPanel extends JPanel implements ComponentListener {
         draw(g);
     }
 
-    private int getCurrentValue() {
+    private long getCurrentValue() {
         return this.currentValue;
     }
